[Refactor] Split main.rs into several modules
* `extractors` contains the actix_web guards
* `handlers` contains all API endpoints
* `app_state` contains the global state of the application
